% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefs1}
\alias{coefs1}
\title{\code{coefs1}}
\format{
A data frame of 14 rows and 6 variables:
\describe{
\item{term}{coefficient name}
\item{estimate}{regression estimate}
\item{p.value}{p value of regression estimate}
\item{rsq}{model r square}
\item{p.mod}{model p value}
\item{check}{direction check}
}
}
\usage{
coefs1
}
\description{
Dataset containing example regression outputs from a key driver
analysis.
}
\keyword{datasets}
