% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labels1}
\alias{labels1}
\title{\code{labels1}}
\format{
A data frame of 8 rows and 2 variables:
\describe{
\item{var}{coefficient name}
\item{label}{label to use in place of var in output}
}
}
\usage{
labels1
}
\description{
Dataset containing example labels used for outputs in key driver
analysis.
}
\keyword{datasets}
