% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{perf1}
\alias{perf1}
\title{\code{perf1}}
\format{
A data frame of 8 rows and 4 variables:
\describe{
\item{Variable}{variable name that matches label var and term prefix}
\item{Negative}{percent negative performance, max = 1 for 100\%}
\item{Neutral}{percent neutral performance, max = 1 for 100\%}
\item{Positive}{percent positive performance, max = 1 for 100\%}
}
}
\usage{
perf1
}
\description{
Dataset containing example performance data from a key driver
analysis.
}
\keyword{datasets}
