% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_coefs}
\alias{process_coefs}
\title{Syntax for processing coefficients captured by broom::tidy for use with
other key driver related functions}
\usage{
process_coefs(DATA, LABELS, DECIMALS, SORT)
}
\description{
Syntax for processing coefficients captured by broom::tidy for use with
other key driver related functions
}
\keyword{internal}
