% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_driver.R
\name{qplot_driver}
\alias{qplot_driver}
\title{qplot_driver}
\usage{
qplot_driver(
  data = NULL,
  labels = NULL,
  output = NULL,
  sort = TRUE,
  legend = TRUE,
  decimals = 1,
  skeleton = FALSE
)
}
\arguments{
\item{data}{Dataframe or tibble containing the output from a modeling process.
Expects two columns minimum: term, estimate.  Examine \code{qplot::coefs1} for
an example file that is generated by \code{qmod::get_driver_coefs}.}

\item{labels}{Dataframe that contains the label names to be used in the graphic
output on the plot rather than the default variable names that may be found in
the dataset.  Expects 2 columns named var and label.  Examine \code{qplot::labels1}
as an example dataframe for this parameter.  Optional.}

\item{output}{Character string containing the complete path and file name of
a png file of the resulting performance plot.  If not provided, the plot will
be shown on the screen instead but may not be optimized for screen viewing.
Optional.}

\item{sort}{Logical indicating of the data should be sorted in descending order
of magnitude (taking into account both positive and negative impact combined).
Default: TRUE}

\item{legend}{A logical value indicating if a legend containing the Labels of
Positive, Neutral, and Negative along with their corresponding colors should
be displayed in the bottom right-hand corner of the plot. Default: TRUE}

\item{decimals}{Numeric value indicating the number of decimals to use for
rounding the displayed numbers in the graph. Default: 1}

\item{skeleton}{A logical value to indicate if the syntax should be printed to
the console.  This is useful if the default selections need tweaking so the
user can see the raw syntax that can be used in place of the function.
Default: FALSE}
}
\value{
Outputs a png file of the driver plot if output path specified,
otherwise generates plot on screen.
}
\description{
Generates a driver graph showing the positive and negative effects
from a key driver model in a tornado-type graph.  Output for Key Driver Analyses.
}
\examples{
\dontrun{
if(interactive()){
 qplot_driver(data=qplot::coefs1, labels=qplot::labels1)
 }
}
}
