% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_nps.R
\name{qplot_nps}
\alias{qplot_nps}
\title{qplot_nps}
\usage{
qplot_nps(
  data,
  var,
  nps,
  var_text = NULL,
  level_labels = NULL,
  output = NULL,
  legend = TRUE,
  rm.na = FALSE,
  sizes = FALSE,
  tukey = FALSE,
  skeleton = FALSE,
  data_out = FALSE,
  ps = 8
)
}
\arguments{
\item{data}{Dataframe name that contains the data to be plotted.  Required.}

\item{var}{The name of the variable plotting variable whose values will form
the groups across which the NPS average will be calculated.}

\item{nps}{The name of the NPS variable in the dataframe, without quotes. Required.}

\item{var_text}{Character string to be used in the plot in place of the name
of the var variable.  If not provided, the variable name will be used.  Optional.}

\item{level_labels}{Character vector of labels to be used along the horizontal access,
one for each value of var.  Do not include an NA label as this will automatically
be created.  If not labels are provided, the values themselves are used.  Optional.}

\item{output}{output Character string containing the complete path and file name of
a png file of the resulting performance plot.  If not provided, the plot will
be shown on the screen instead but may not be optimized for screen viewing.
Optional.}

\item{legend}{A logical value indicating if a legend containing the Labels of
Positive, Neutral, and Negative along with their corresponding colors should
be displayed in the bottom right-hand corner of the plot. Default: TRUE}

\item{rm.na}{A logical value indicating if missing values should be removed
from var or not.  Default: FALSE}

\item{sizes}{A logical value indicating if the frequency and percentage of each
category of var along the horizontal access is shown with the level labels on
the plot.  Default: FALSE}

\item{tukey}{A logical value indicating if an ANOVA and Tukey Tests for contrasts
should be performed}

\item{skeleton}{A logical value to indicate if the syntax should be printed to
the console.  This is useful if the default selections need tweaking so the
user can see the raw syntax that can be used in place of the function.
Default: FALSE}

\item{data_out}{Logical indicating if the plot data should be returned.  If
tukey parameter is set to TRUE, then it will output a list containing both the
plot data and the tukey data.}

\item{ps}{Point size setting for grDevices::jpeg.  May need to reduce if lots
of attributes are included.  Default=8.}
}
\value{
Outputs a png file of the nps plot if output path specified,
otherwise generates plot on screen.
}
\description{
Generates a barchart of the average NPS score for every value of a
second variable, likely a key driver prospect.
}
\examples{
\dontrun{
if(interactive()){
 qplot::nps1 \%>\%
 dplyr::select(-response_id) \%>\%
   qplot_nps(data=., var=price, nps=nps, var_text="Price", rm.na=TRUE,
             level_labels=c("Very Unsatisifed", "Unsatisfied", "Neither", "Satisified", "Very Satisfied"))
 }
}
}
