% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_performance.R
\name{qplot_performance}
\alias{qplot_performance}
\title{qplot_performance}
\usage{
qplot_performance(
  data = NULL,
  labels = NULL,
  order = NULL,
  output = NULL,
  legend = TRUE,
  decimals = 1,
  skeleton = FALSE,
  data_out = FALSE,
  ps = 10
)
}
\arguments{
\item{data}{Dataframe name that contains the performance data.  Expects 4 columns
named Variable, Negative, Neutral, and Positive.  Variable is the name of each
attribute in truncated form, similar to a coefficient name.  Examine \code{qplot::perf1}
as an example dataframe for this parameter.  Required.}

\item{labels}{Dataframe that contains the label names to be used in the graphic
output on the plot rather than the default variable names that may be found in
the dataset.  Expects 2 columns named var and label.  Examine \code{qplot::labels1}
as an example dataframe for this parameter.  Optional.}

\item{order}{Dataframe that contains the tidied output from a modeling process
like regression that contains the coefficients, estimates, p values, etc. Expects
columns named term and estimate, although such files often contain other columns.
See \code{qplot::coefs1} for an example dataframe for this parameter. The file will
be processed to put the coefficients in decreasing order of magnitude.  The
term prefixes should match the var column (from labels) and Variable (from data).
The Performance Graph will also be sorted according to the same decreasing
order of magnitude as the estimates. Optional.}

\item{output}{Character string containing the complete path and file name of
a png file of the resulting performance plot.  If not provided, the plot will
be shown on the screen instead but may not be optimized for screen viewing.
Optional.}

\item{legend}{A logical value indicating if a legend containing the Labels of
Positive, Neutral, and Negative along with their corresponding colors should
be displayed in the bottom right-hand corner of the plot. Default: TRUE}

\item{decimals}{Numeric value indicating the number of decimals to use for
rounding the displayed numbers in the graph. Default: 1}

\item{skeleton}{A logical value to indicate if the syntax should be printed to
the console.  This is useful if the default selections need tweaking so the
user can see the raw syntax that can be used in place of the function.
Default: FALSE}

\item{data_out}{A logical value to indicate if the process data file should be
returned when skeleton=TRUE.  Default: FALSE}

\item{ps}{Point size setting for grDevices::jpeg.  May need to reduce if lots
of attributes are included.  Default=10.}
}
\value{
Outputs a png file of the performance plot if output path specified,
otherwise generates plot on screen.
}
\description{
Generates a performance graph showing Positive, Neutral, and Negative
proportions.  Output for Key Driver Analyses.
}
\examples{
\dontrun{
if(interactive()){
 qplot_performance(data=qplot::perf1, labels=qplot::labels1,
                   output="./output/test1.png")
 }
}
}
