% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refresh}
\alias{refresh}
\title{Refreshes github install}
\usage{
refresh(dev = FALSE)
}
\arguments{
\item{dev}{A logical to indicate if the development version of qplot should be
installed.  Will only work if qplot is the only package listed in the pack_list.
Default: FALSE.}
}
\value{
Nothing, updates package(s).
}
\description{
Convenience function to perform a remotes::install_github for qplot.
}
\examples{
\dontrun{
qplot:::refresh()
qplot:::refresh(dev=TRUE)
}
}
\keyword{internal}
