% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test1}
\alias{test1}
\title{test1 is an example dataset containing labelled variables for testing}
\format{
A data frame of 200 observations and 11 variables:
\describe{
\item{response_id}{Respondent ID Number}
\item{q1}{Age Category}
\item{q2}{Gender}
\item{q3}{Race}
\item{q4}{Type of Insurance}
\item{q5}{Income}
\item{q6}{Education}
\item{region}{Region}
\item{prob1}{Problem - No Variable Label}
\item{prob2}{Problem - Character Data}
\item{prob3}{Problem - No Value labels}
}
}
\usage{
test1
}
\description{
Dataset containing a few variables with labelled data, as well
as three problem variables for detecting issues with labelled data.
}
\keyword{datasets}
